refactor(shell): extract deep link routing controller
